Cast `len` to u32

+++ parachain/test-parachains/adder/src/wasm_validation.rs
@@ -24,7 +24,7 @@ use codec::{Encode, Decode};
 
 #[no_mangle]
 pub extern "C" fn validate_block(params: *const u8, len: u32) -> u64 {
-	let params = unsafe { parachain::load_params(params, len) };
+	let params = unsafe { parachain::load_params(params, len as u32) };
 	let parent_head = HeadData::decode(&mut &params.parent_head.0[..])
 		.expect("invalid parent head format.");
